requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pyjwt